migrations-ee: rename module
Renamed module to 'migrations-ee'.
