+++ deps.sh
@@ -7,4 +7,4 @@ set -e
 tarantoolctl rocks install luatest 0.5.7
 tarantoolctl rocks install luacov 0.13.0
 tarantoolctl rocks install luacheck 0.26.0
-tarantoolctl rocks make migrations-scm-1.rockspec
+tarantoolctl rocks make migrations-ee-scm-1.rockspec
