+++ test/integration/basic_test.lua
@@ -56,8 +56,8 @@ local cases = {
     with_config_loader = function()
         for _, server in pairs(g.cluster.servers) do
             server.net_box:eval([[
-                require('migrator').set_loader(
-                    require('migrator.config-loader').new()
+                require('migrator-ee').set_loader(
+                    require('migrator-ee.config-loader').new()
                 )
             ]])
         end
@@ -73,8 +73,8 @@ local cases = {
     with_directory_loader = function()
         for _, server in pairs(g.cluster.servers) do
             server.net_box:eval([[
-                require('migrator').set_loader(
-                    require('migrator.directory-loader').new('test/integration/migrations')
+                require('migrator-ee').set_loader(
+                    require('migrator-ee.directory-loader').new('test/integration/migrations')
                 )
             ]])
         end
@@ -208,18 +208,18 @@ g.test_gh_66_configurable_timeout = function(cg)
     local main = g.cluster.main_server
 
     main:eval([[
-        require('cartridge').config_patch_clusterwide({migrations = {applied = {}, options = {storage_timeout = 0.1}}})
+        require('cartridge').config_patch_clusterwide({['migrations-ee'] = {applied = {}, options = {storage_timeout = 0.1}}})
     ]])
 
     for _, server in pairs(cg.cluster.servers) do
         server.net_box:eval([[
-            require('migrator').set_loader(
-                require('migrator.directory-loader').new('test/integration/migrations-gh-66')
+            require('migrator-ee').set_loader(
+                require('migrator-ee.directory-loader').new('test/integration/migrations-gh-66')
             )
         ]])
     end
 
-    local status, resp = g.cluster.main_server:eval("return pcall(function() require('migrator').up() end)")
+    local status, resp = g.cluster.main_server:eval("return pcall(function() require('migrator-ee').up() end)")
     t.assert_equals(status, false)
     t.assert_str_contains(tostring(resp), 'Errors happened during migrations')
