+++ test/integration/migrations_check_roles_enabled/01_first.lua
@@ -1,4 +1,4 @@
-local utils = require('migrator.utils')
+local utils = require('migrator-ee.utils')
 
 return {
     up = function()
